include/wb.h: Avoid duplicate calculation.
Since we are going to compute the product of two variables anyway and store
it in a variable, move its computation earlier and use it to avoid computing
the multiplication more than once.

Also, we fix the type of the variable so that we don't get warnings by the
compiler telling us that we are comparing variables of different types.